feat: ProductCard 컴포넌트 분리 및 추가기능
- width를 props로 받아 변경할 수 있도록 함
